Extracted functional test
