+++ test/construct_ballot_cryptograms.test.ts
@@ -3,8 +3,6 @@ import { expect } from 'chai';
 import nock = require('nock');
 import { deterministicRandomWords, deterministicMathRandom, resetDeterministicOffset } from './test_helpers';
 import sinon = require('sinon');
-import { validateCvr } from '../lib/av_client/cvr_validation';
-import { Option } from '../lib/av_client/types';
 const sjcl = require('../lib/av_client/sjcl')
 
 describe('AVClient#constructBallotCryptograms', () => {
@@ -57,41 +55,35 @@ describe('AVClient#constructBallotCryptograms', () => {
     });
   });
 
-  const createOptions = (optionHandles: string[]): Option[] => {
-    return optionHandles.map((handle, i) => {
-      return {
-        "id": i,
-        "handle": handle,
-        "title": {},
-        "subtitle": {},
-        "description": {},
-        "url": {},
-        "video_url": {},
-        "image": null,
-        "ancestry": null,
-        "selectable": true
-      }
-    })
-  }
-
-
   context('given invalid CVR', () => {
-    it('encryption fails when voting on invalid contest / encryption fails when voting on invalid option', async () => {
-      const template = { title: {}, vote_encoding_type: 0, description: {}, write_in: false }
+    it('encryption fails when voting on invalid contest', async () => {
+      await client.requestAccessCode('voter123', 'voter@foo.bar');
+      await client.validateAccessCode('1234');
+      await client.registerVoter()
 
-      const contest1 = { ...template, id: 1, options: createOptions(['option1', 'option2']) };
-      const contest2 = { ...template, id: 2, options: createOptions(['optionA', 'optionB']) };
-      const contest3 = { ...template, id: 3, options: createOptions(['optionX', 'optionY']) };
+      const cvr = { '1': 'option1', '3': 'optiona' };
 
-      const ballot1 = [contest1, contest2];
-      const ballot2 = [contest1, contest3];
+      try {
+        await client.constructBallotCryptograms(cvr);
+        expect.fail('Expected an error to be thrown');
+      } catch(error) {
+        expect(error.message).to.equal('Corrupt CVR: Contains invalid contest');
+      }
+    });
 
-      const cvr1 = { '1': 'option1', '3': 'optionX' };
-      const cvr2 = { '1': 'option1', '3': 'optionINVALID' };
+    it('encryption fails when voting on invalid option', async () => {
+      await client.requestAccessCode('voter123', 'voter@foo.bar');
+      await client.validateAccessCode('1234');
+      await client.registerVoter()
 
-      expect(validateCvr(cvr2, ballot2)).to.be.equal(':invalid_option')
-      expect(validateCvr(cvr2, ballot1)).to.be.equal(':invalid_contest')
-      expect(validateCvr(cvr1, ballot2)).to.be.equal(':okay')
+      const cvr = { '1': 'option1', '2': 'wrong_option' };
+
+      try {
+        await client.constructBallotCryptograms(cvr);
+        expect.fail('Expected error to be thrown');
+      } catch(error) {
+        expect(error.message).to.equal('Corrupt CVR: Contains invalid option');
+      }
     });
   });
 });
